+++ EDDI/ChangeLog.md
@@ -11,6 +11,7 @@ Full details of the variables available for each noted event, and VoiceAttack in
     * Added new event `Statistics`, triggered while loading a game.
     * Added new event `Powerplay`, triggered while loading the game (if pledged).
     * Updated the `Commander continued` event with new properties `startlanded` and `startdead` (true if starting the game either landed or dead, respectively).
+    * Updated the `Community goal` event with new properties `maxtier` and `maxtierrewards`.
     * Updated the `System report` script to enhance the description of powerplay status.
     * Updated the `Engineer progressed` event to stay silent by default when written at startup (with empty values, signaling that engineer data has been loaded).
     * Updated the `Bodies mapped` script to correct a typo.